requirements.txt
requests
selenium
beautifulsoup4
undetected-chromedriver
proxytools
selenium-stealth
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-docx
dotenv
markdown
beautifulsoup4
mistune